# Improve pbf parser performance