--- a/core/src/main/java/com/graphhopper/reader/ReaderNode.java
+++ b/core/src/main/java/com/graphhopper/reader/ReaderNode.java
@@ -34,8 +34,6 @@ public class ReaderNode extends ReaderElement {
 
         this.lat = pointAccess.getLatitude(accessId);
         this.lon = pointAccess.getLongitude(accessId);
-        if (pointAccess.is3D())
-            setTag("ele", pointAccess.getElevation(accessId));
     }
 
     public ReaderNode(long id, double lat, double lon) {
@@ -51,36 +49,6 @@ public class ReaderNode extends ReaderElement {
 
     public double getLon() {
         return lon;
-    }
-
-    public double getEle() {
-        Object ele = getTags().get("ele");
-        if (ele == null)
-            return Double.NaN;
-        return (Double) ele;
-    }
-
-    @Override
-    public void setTag(String name, Object value) {
-        if ("ele".equals(name)) {
-            if (value == null)
-                value = null;
-            else if (value instanceof String) {
-                String str = (String) value;
-                str = str.trim().replaceAll("\\,", ".");
-                if (str.isEmpty())
-                    value = null;
-                else
-                    try {
-                        value = Double.parseDouble(str);
-                    } catch (NumberFormatException ex) {
-                        return;
-                    }
-            } else
-                // force cast
-                value = ((Number) value).doubleValue();
-        }
-        super.setTag(name, value);
     }
 
     @Override