--- a/core/src/test/java/com/graphhopper/reader/OSMElementTest.java
+++ b/core/src/test/java/com/graphhopper/reader/OSMElementTest.java
@@ -40,7 +40,7 @@ public class OSMElementTest {
     @Test
     public void testSetTags() {
         ReaderElement instance = new ReaderWay(1);
-        Map<String, String> map = new HashMap<>();
+        Map<String, Object> map = new HashMap<>();
         map.put("test", "xy");
         instance.setTags(map);
         assertTrue(instance.hasTag("test", "xy"));