… core/src/test/java/com/graphhopper/reader/OSMNodeTest.java
… reader-osm/src/test/resources/com/graphhopper/reader/osm/custom-osm-ele.xml