This should speedup pbf parsing by about 10%