@otbutz 2020-11-30
#2191 reduced parsing duration for my test dataset from 74s to 50s. With the changes of this PR we're down to about 38s.