@otbutz 2020-12-01
Fixes #2197