@otbutz 2020-12-08
Most of the commits aim to reduce object creation and collection resizes which are both costly operations if they're often repeated.

https://github.com/graphhopper/graphhopper/pull/2193/commits/0f376c6de76cfe2b6be1ff2541490410cbafb8d2 tries to avoid additional lookups in the tree by passing through internal IDs from previous lookups.

https://github.com/graphhopper/graphhopper/pull/2193/commits/16eb7166cd93a89491f1e55661aaba4991ac87e7 follows the OSM pbf spec as we only need to process one typ of object within a PrimitiveGroup.