@otbutz 2020-12-08
Is there any reason we keep ReaderNode.getEle() ? It's an ugly hack within setTag and isn't really used for routing.