@otbutz 2020-12-08
@karussell i removed the node skipping logic as it was quite an invasive change with little benefit. Also removed the elevation handling from the reader.