@otbutz 2020-12-09
@karussell any remarks? I'd like to get this merged as my other parsing related PRs are based on this.