@easbar 2020-12-11
I did not look at the code yet, but I can confirm that this branch is faster and the numbers of nodes and edges stays the same. For germany-190101.osm.pbf I got:

```
#master:
com.graphhopper.reader.osm.OSMReader - time pass1:114s, pass2:198s, total:312s
#this branch:
com.graphhopper.reader.osm.OSMReader - time pass1:109s, pass2:178s, total:288s
```