@otbutz 2020-12-11
Thanks for testing :smiley: 

https://github.com/graphhopper/graphhopper/pull/2205 should yield even better result while improving the readability of the code