@karussell 2020-12-11
In general it is not clear which change is only a refactoring or a clean up and which is to improve performance - can you move the unrelated changes into a separate PR?

> 0f376c6 tries to avoid additional lookups in the tree by passing through internal IDs from previous lookups.

The OSMReader is already a complicated mess ;) so I would try to avoid making it worse. And the "graphIndex" would add to this. Is this required for the speed up? I would really like to clean up and/or refactor OSMReader to make it faster, but readability and maintainability should be a higher priority now that reading PBFs is acceptable fast (and usually only a small part when CH or LM is involved).

> 16eb716 follows the OSM pbf spec as we only need to process one typ of object within a PrimitiveGroup.

Is this required for the speed up?

> #2205 should yield even better results while improving the readability of the code

What is the reason you drop support for XML in #2205? And if this is done it should be done in a separate PR. In general smaller PRs can be merged faster :)