@otbutz 2020-12-11
> In general it is not clear which change is only for refactoring (or clean up) and which is to improve performance - can you move the unrelated changes into a separate PR?

Sure

> The OSMReader is already a complicated mess ;)

Agreed :sweat_smile: 

> so I would try to avoid making it worse. And the "graphIndex" would add to this. Is this required for the speed up?

See #2194 for more details. It was an attempt to reduce the number of lookups.

> I would really like to clean up and/or refactor OSMReader to make it faster, but readability and maintainability should be a higher priority now that reading PBFs is acceptable fast (and usually only a small part when CH or LM is involved).

Fair point. I'm going to remove it.

> > [16eb716](https://github.com/graphhopper/graphhopper/commit/16eb7166cd93a89491f1e55661aaba4991ac87e7) follows the OSM pbf spec as we only need to process one typ of object within a PrimitiveGroup.
> 
> Is this required for the speed up?

Just a minor bugfix. I will move it to a separate PR.

> What is the reason you drop support for XML in #2205?

The current implementation has to deal with both formats which leaves little room for optimizations and makes the code less readable. e.g osm2pgsql requires all imported pbfs to be [sorted by type and id](https://github.com/openstreetmap/osm2pgsql/releases/tag/1.4.0) now. If we follow their requirement it could help us to improve #2194 

See also our discussion in https://github.com/graphhopper/graphhopper/issues/2130

Anybody who wants to import XML files can still convert them with tools like osmconvert or osmium. 

> And if this is done it should be done in a separate PR. In general smaller PRs can be merged faster :)

It's in fact already a separate branch but i haven't created PR.