@easbar 2020-12-13
I ran master vs this branch a few more times:

```sh
# master
2020-12-13 15:48:08.522 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:180s, pass2:280s, total:460s
2020-12-13 15:57:45.614 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:178s, pass2:283s, total:462s
2020-12-13 16:07:44.938 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:187s, pass2:296s, total:483s
2020-12-13 16:17:48.119 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:191s, pass2:292s, total:483s
2020-12-13 16:27:18.800 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:175s, pass2:279s, total:455s
2020-12-13 16:37:04.125 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:181s, pass2:287s, total:469s
2020-12-13 16:46:39.075 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:280s, total:459s
2020-12-13 16:56:11.371 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:276s, total:455s
2020-12-13 17:05:46.935 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:281s, total:461s
2020-12-13 17:15:35.966 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:189s, pass2:282s, total:471s
2020-12-13 17:25:14.669 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:281s, total:460s
2020-12-13 17:34:50.932 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:278s, total:458s
# speedup_pbf      
2020-12-13 09:48:14.702 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:170s, pass2:264s, total:434s
2020-12-13 09:57:27.834 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:170s, pass2:268s, total:438s
2020-12-13 10:06:51.245 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:179s, pass2:268s, total:447s
2020-12-13 10:16:08.205 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:173s, pass2:268s, total:441s
2020-12-13 10:25:36.795 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:177s, pass2:273s, total:451s
2020-12-13 10:34:49.561 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:170s, pass2:267s, total:437s
2020-12-13 10:44:07.813 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:174s, pass2:268s, total:442s
2020-12-13 10:53:29.824 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:172s, pass2:272s, total:445s
2020-12-13 11:02:47.947 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:172s, pass2:269s, total:441s
2020-12-13 11:12:28.436 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:182s, pass2:281s, total:463s
2020-12-13 11:21:45.686 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:171s, pass2:270s, total:441s
2020-12-13 11:31:05.948 [main] INFO  com.graphhopper.reader.osm.OSMReader - time pass1:174s, pass2:270s, total:444s
```

The speedup seems like less than 10%? I was trying to check whether all the commits yield an improvement or if some are more crucial than others, but this wasn't very clear, because the differences were rather small.